requirements.txt
google-cloud-aiplatform>=1.106.0
google-generativeai>=0.8.3
google-genai>=0.4.0
langchain>=0.3.0
langchain-google-vertexai>=2.0.27
langchain-community>=0.3.0
google-cloud-storage==2.18.0
google-cloud-functions>=1.18.0
pydantic>=2.10.3
pydantic-settings>=2.7.0
google-cloud-secret-manager>=2.24.0
google-cloud-firestore>=2.20.0
google-cloud-pubsub>=2.25.0
google-cloud-discoveryengine==0.11.1
google-cloud-aiplatform>=1.106.0
pandas
python-dotenv
google-auth>=2.36.0
google-auth-oauthlib>=1.2.2
google-api-python-client>=2.156.0
functions-framework>=3.9.2
python-magic>=0.4.27
requests>=2.32.3
httpx>=0.28.1
aiohttp>=3.11.7
python-dotenv>=1.0.1
pydantic>=2.10.3
pydantic-settings>=2.7.0
pypdf>=5.1.0
docx2txt>=0.8
python-docx>=1.1.2
Pillow>=11.0.0
structlog>=24.5.0
rich>=13.9.4
pytest>=8.3.4
pytest-asyncio>=0.25.0
black>=24.10.0
flask>=3.1.1
pdf2image==1.17.0
protobuf~=5.29.5
asyncio~=3.4.3